Authorization

+++ dotnet-rpg/Services/CharacterService.cs
@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_rpg.Data;
 using dotnet_rpg.Dtos.Character;
 using dotnet_rpg.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace dotnet_rpg.Services
@@ -13,16 +15,24 @@ namespace dotnet_rpg.Services
     public class CharacterService : ICharacterService
     {
         private readonly IMapper _mapper;
+
         private readonly DataContext _context;
-        public CharacterService(IMapper mapper, DataContext context)
+
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
-            public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+      
+        public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character= _mapper.Map<Character>(newCharacter);
+            character.User =await _context.Users.FirstOrDefaultAsync(u=> u.Id == GetUserId());
            _context.Characters.Add(character);
             await _context.SaveChangesAsync();  
             serviceResponse.Data= await _context.Characters.Select(c=> _mapper.Map<GetCharacterDto>(c)).ToListAsync();
@@ -34,10 +44,20 @@ namespace dotnet_rpg.Services
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character =await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                Character character = await _context.Characters.FirstAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _context.Characters
+                        .Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Characters not found!";
+                }
             }
             catch (Exception ex)
             {
@@ -50,14 +70,14 @@ namespace dotnet_rpg.Services
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _context.Characters.ToListAsync();
+            var dbCharacters = await _context.Characters.Where(c=> c.User.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return serviceResponse;
         }
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacters = await _context.Characters.FirstOrDefaultAsync(c=> c.Id == id);
+            var dbCharacters = await _context.Characters.FirstOrDefaultAsync(c=> c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacters);
             return serviceResponse;
         }
@@ -67,17 +87,25 @@ namespace dotnet_rpg.Services
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
             try
             {
-                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                Character character = await _context.Characters
+                    .Include(c=>c.User)
+                    .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                if (character.User.Id == GetUserId())
+                {
+                    character.Name = updatedCharacter.Name;
+                    character.HitPoints = updatedCharacter.HitPoints;
+                    character.Strenght = updatedCharacter.Strenght;
+                    character.Defense = updatedCharacter.Defense;
+                    character.Class = updatedCharacter.Class;
+                    character.Intelligence = updatedCharacter.Intelligence;
 
-                character.Name = updatedCharacter.Name;
-                character.HitPoints = updatedCharacter.HitPoints;
-                character.Strenght = updatedCharacter.Strenght;
-                character.Defense = updatedCharacter.Defense;
-                character.Class = updatedCharacter.Class;
-                character.Intelligence = updatedCharacter.Intelligence;
-
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+                }
+                else {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Characters not found!";
+                }
             }
             catch (Exception ex)
             {
